refactor(Stream): remove content/ directory
