+++ src/component/container/AppMainColumn/GitHubUserEventContainer/GitHubUserEventContainerCommandBarContainer/GitHubUserEventContainerCommandBarContainer.tsx
@@ -4,7 +4,7 @@ import { IconButton } from "office-ui-fabric-react";
 import { BaseContainer } from "../../../BaseContainer";
 import classNames from "classnames";
 import { ExpandableSearch } from "../../../../project/ExpandableSearch/ExpandableSearch";
-import { createApplyFilterToCurrentStreamUseCase } from "../../../../../use-case/content/ApplyFilterToCurrentStreamUseCase";
+import { createApplyFilterToCurrentStreamUseCase } from "../../../../../use-case/GitHubSearchStream/ApplyFilterToCurrentStreamUseCase";
 import { createReloadCurrentUserActivityUseCase } from "../../../../../use-case/GitHubUser/ReloadCurrentUserActivityUseCase";
 
 export interface GitHubUserEventContainerCommandBarContainerProps {
