+++ src/use-case/GitHubSearchStream/ApplyFilterToCurrentStreamUseCase.ts
