--- src/use-case/content/README.md
@@ -1,6 +1,0 @@
-# Content
-
-Main contend is either one of following:
-
-- GitHubStream
-- GitHubUserActivity
